feat: default content for search component
