+++ packages/component-library/src/SearchButton/index.tsx
@@ -3,10 +3,13 @@
 import { MagnifyingGlass } from "@phosphor-icons/react/dist/ssr/MagnifyingGlass";
 import clsx from "clsx";
 import type { ReactNode } from "react";
+import { useMemo } from "react";
+import { useIsSSR } from "react-aria";
 import { Button as BaseButton } from "react-aria-components";
 
 import type { Props as ButtonProps } from "../Button";
 import { Button } from "../Button";
+import { Skeleton } from "../Skeleton";
 import styles from "./index.module.scss";
 
 type OwnProps = {
@@ -19,6 +22,16 @@ type Props = Pick<
   "beforeIcon" | "size" | "onClick" | "className" | "isPending"
 > &
   OwnProps;
+
+const SearchShortcutText = () => {
+  const isSSR = useIsSSR();
+  return isSSR ? <Skeleton width={7} /> : <SearchTextImpl />;
+};
+
+const SearchTextImpl = () => {
+  const isMac = useMemo(() => navigator.userAgent.includes("Mac"), []);
+  return isMac ? "⌘ K" : "Ctrl K";
+};
 
 export const SearchButton = ({
   beforeIcon,
@@ -36,7 +49,7 @@ export const SearchButton = ({
         rounded
         {...props}
       >
-        {largeScreenContent}
+        {largeScreenContent ?? <SearchShortcutText />}
       </Button>
       <Button
         className={clsx(styles.smallScreenSearchButton, props.className)}
@@ -47,7 +60,7 @@ export const SearchButton = ({
         rounded
         {...props}
       >
-        {smallScreenContent}
+        {smallScreenContent ?? <SearchShortcutText />}
       </Button>
     </div>
   );
